Add default RTEMS version support, environment var checking and CC version message.
Allow an application the ability to set a version number for RTEMS. This avoids
issues with the automatic detection code. It means an application becomes keyed
to a specific version of RTEMS.

Check the environment for variables being set that could effect a build. We
allow the environment to do this but it can have a side effect such as CC being
set for one architecture and the rtems_waf being asked to use another.

Print the version of CC being used. This is a diagnostic.