Allow vendor field in toolchain target triplet
Rework the splitting of arch and bsp to rely on the last field in the
arch section starting with "rtems", instead of relying on the arch being
exactly two fields in size.

This makes sure that toolchains with a vendor field in their target
triplet can be used with this build system.

Toolchains produced by the RTEMS source builder tend to omit the vendor
field, but for example the SPARC LEON/ERC32 toolchain provided by
Gaisler through the RCC package does include a vendor field.